requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
WTForms==3.0.1
PyMySQL==1.1.0
pandas>=2.0.0,<2.2.0
numpy>=1.24.0,<1.26.0
matplotlib==3.7.2
seaborn==0.12.2
wordcloud==1.9.2
jieba==0.42.1
requests==2.31.0
python-dotenv==1.0.0
Werkzeug==2.3.7
Jinja2==3.1.2
beautifulsoup4==4.12.2
lxml==4.9.3
kaggle==1.5.16
